NEWLY MODIFIED ADDPROJECT
